Add ability to read in implany from file
This is important for usability, since all implant will be written as a
file.

To account for the multi-line aspect, had to rework the execution code
in the receiver.

As a result of reworking the execution pipe, no longer needed the single
quotes on each side of the implant, so removed those.